implementing User versioning and content negotiation for API

+++ Cab-Booking-System/Cab-Booking-System/src/main/java/com/RideMate/Cab_Booking_System/Services/UserService.java
@@ -5,6 +5,7 @@ import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
+import com.RideMate.Cab_Booking_System.Controllers.UserController.UserNotFoundException;
 import com.RideMate.Cab_Booking_System.Entities.User;
 import com.RideMate.Cab_Booking_System.Repositories.UserRepository;
 
@@ -25,19 +26,34 @@ public class UserService {
 		return repository.findAll();
 	}
 	
-	public Optional<User> getUserById(Integer id) {
+	public User getUserById(Integer id) {
 		Optional<User> existingUser = repository.findById(id);
-		return existingUser;
+		
+		if(existingUser.isEmpty()) {
+			throw new UserNotFoundException("User not found by id -> " + id);
+		}
+		
+		return existingUser.get();
 	}
 	
 	public void updateExistingUser(User user, Integer id) {
-		User existingUser = repository.findById(id).get();
+		User existingUser = repository.findById(id)
+				.orElseThrow(()-> new UserNotFoundException("User not Found by id -> " + id));
+				
 		existingUser.setName(user.getName());
 		existingUser.setPhone(user.getPhone());
-		existingUser
+		
+		repository.save(existingUser);
 	}
 	
 	public void deleteExistingUser(Integer id) {
+		
+		Optional<User> existingUser = repository.findById(id);
+		
+		if(existingUser.isEmpty()) {
+			throw new UserNotFoundException("user not found by id -> " + id);
+		}
+		
 		repository.deleteById(id);
 	}
